requirements.txt
selenium==4.23.1
mysql-connector-python==9.1.0
pandas==2.3.2
webdriver-manager
seaborn
matplotlib
streamlit